fix: Monitoring Dashboard Workloads - Controllers List: List 컴포넌트에서 데이터 로드 하도록 수정
